refactor: move `pluralize` into its own module
