+++ extensions/ql-vscode/src/pure/word.ts
@@ -1,0 +1,8 @@
+/**
+ * Pluralizes a word.
+ * Example: Returns "N repository" if N is one, "N repositories" otherwise.
+ */
+
+export function pluralize(numItems: number | undefined, singular: string, plural: string): string {
+  return numItems !== undefined ? `${numItems} ${numItems === 1 ? singular : plural}` : '';
+}
