+++ extensions/ql-vscode/src/remote-queries/run-remote-query.ts
@@ -22,7 +22,8 @@ import { OctokitResponse, RequestError } from '@octokit/types/dist-types';
 import { RemoteQuery } from './remote-query';
 import { RemoteQuerySubmissionResult } from './remote-query-submission-result';
 import { QueryMetadata } from '../pure/interface-types';
-import { getErrorMessage, REPO_REGEX, pluralize } from '../pure/helpers-pure';
+import { getErrorMessage, REPO_REGEX } from '../pure/helpers-pure';
+import { pluralize } from '../pure/word';
 import * as ghApiClient from './gh-api/gh-api-client';
 import { getRepositorySelection, isValidSelection, RepositorySelection } from './repository-selection';
 import { parseVariantAnalysisQueryLanguage, VariantAnalysisSubmission } from './shared/variant-analysis';
